More Code sorting
just sorting code out to be more manageable.
