+++ TopDownGame-Var3/main.lua
@@ -14,6 +14,7 @@ function love.load() -- this function loads everything when game starts
     require ('dabrain/bullets')
     require ('player')
     require('dabrain/monsterTimer')
+    require('dabrain/monsterSpawner')
 end
 
 function love.update(dt) -- this is the "game loop" that runs at 60FPS
@@ -51,7 +52,7 @@ function love.draw() -- this function handles drawing the graphics.
     -- this draws the crosshair sprite, and assigns the position to the mouseX and mouseY position.
     --  also, you need to subtract the PNG size to correct for the offset position of the PNG
     love.graphics.draw(sprites.crosshairs, love.mouse.getX()-20, love.mouse.getY()-20, nil, .75, .75)   
-end
+end --End draw function
 
 
 function love.keypressed(key)
@@ -101,95 +102,8 @@ function zombiePlayerAngle(enemy)
     return math.atan2(player.y - enemy.y, player.x - enemy.x)
 end
 
---[[
-    This function creates the zombie and assigns the zombie to the global "zombies" table.
-]]
-function monsterSelector()
-    monsterSelection = math.random(1,2)
-    if monsterSelection == 1 then
-        spawnZombie()
-    else
-        spawnTarget()
-    end
-end
 
-    function spawnZombie()
-        local zombie = {} -- creates local zombie variable
-        zombie.x = 0 -- creates x position variable
-        zombie.y = 0 -- creates y position variable
-        zombie.speed = 140 -- assigns the zombie speed
-        zombie.dead = false -- when zombie collides with bullet, dead will be set to true and the zombie is removed.
 
-        --[[
-            **RANDOM ZOMBIE PLACEMENT**
-            Goal here is to have zombies walk in from random spot off the screen.
-            side variable gets a random number 1-4 which represents each side of the screen.
-            1 = left, 2 = right, 3 = top, 4 = bottom
-
-        ]]
-        local side = math.random(1,4)
-        if side == 1 then
-            zombie.x = -30
-            zombie.y = math.random(0, love.graphics.getHeight())
-        elseif side == 2 then
-            zombie.x = love.graphics.getWidth() + 30
-            zombie.y = math.random(0, love.graphics.getHeight())
-        elseif side == 3 then
-            zombie.x = math.random(0, love.graphics.getWidth())
-            zombie.y = -30
-        elseif side == 4 then
-            zombie.x = math.random(0, love.graphics.getWidth())
-            zombie.y = love.graphics.getHeight() + 30
-        end
-
-        table.insert(zombies, zombie) -- adds the local "zombie" to the global "zombies" table.
-    end
-
-    function spawnTarget()
-        local target = {} -- creates local zombie variable
-        target.x = 0 -- creates x position variable
-        target.y = 0 -- creates y position variable
-        target.speed = 140 -- assigns the zombie speed
-        target.dead = false -- when zombie collides with bullet, dead will be set to true and the zombie is removed.
-        target.injured = false
-
-        --[[
-            **RANDOM ZOMBIE PLACEMENT**
-            Goal here is to have zombies walk in from random spot off the screen.
-            side variable gets a random number 1-4 which represents each side of the screen.
-            1 = left, 2 = right, 3 = top, 4 = bottom
-
-        ]]
-        local side = math.random(1,4)
-        if side == 1 then
-            target.x = -30
-            target.y = math.random(0, love.graphics.getHeight())
-        elseif side == 2 then
-            target.x = love.graphics.getWidth() + 30
-            target.y = math.random(0, love.graphics.getHeight())
-        elseif side == 3 then
-            target.x = math.random(0, love.graphics.getWidth())
-            target.y = -30
-        elseif side == 4 then
-            target.x = math.random(0, love.graphics.getWidth())
-            target.y = love.graphics.getHeight() + 30
-        end
-
-        table.insert(targets, target) -- adds the local "zombie" to the global "zombies" table.
-    end
-
---[[
-    this function creates the bullet and assigns the bullet to the global "bullets" table
-]]
-function spawnBullet()
-    local bullet = {} -- creates local bullets table
-    bullet.x = player.x -- assigns position x to player location
-    bullet.y = player.y -- assigns position y to player location 
-    bullet.speed = 500 -- assings the bullet speed
-    bullet.direction = playerMouseAngle() -- uses the player mouse angle for the direction of the bullet will go
-    bullet.dead = false -- when bullets collide with zombies, dead will be set to true and bullet is removed
-    table.insert(bullets, bullet) -- inserts bullet into global bullets table.
-end
 
 --[[
     this is the distance between formula that we used in the previous lesson. This simply calculates the distance between two objects.
